Logging update and Reach-with-Lake compatibilty
Updates were made to the logging, such that all print statements will trigger both an arcpy.AddMessage call and a print, with all printed statements being sent to the log file.

Additionally, the capability was added, through the LAKE_PROCESSING_MODULE script, to perform reach-based lake pre-processing in order to support WRF-Hydro (Community) in a reaches with lakes configuration.