Change default lake ID value, lake parameter values
This change removes the requirement of a field "NEWID" in the input lake feature class. The field "NEWID" should be added if it does not exist already in the input, but that was not the case until this update was applied. Also, orifice and weir elevation dictoinaries were updated to include lakes that could not be resolved on the routing grid.