Restoring custom geotransformation enhancements
Some changes here for checking the semi-major and semi-minor axes in against each other for both high-res DEM and model grid. Previously, they were compared, but it is also possible that both are set to 0.0, resulting in the erroneous assumption that the datums were identical. Ensuring that they are not 0.0 after checking for equality should help this situation.