Addressing a numpy masked array issue for netCDF
Although a fix was recently added to check the netCDF4 library version and set_auto_mask to False for certain versions, an issue appeared that certain arrays were being masked, possibly due to nodata values being present in the netCDF data variables. A check is done on the latitude and longitude arrays to unmask them if necessary.