Reversing some recently added datum-checking functionality
Unfortunately, a check that was put in recently to compare datums between the model coordinate system and the input high-resolution DEM are not working properly because of a potential bug in Esri spatialreference objects. The datum information comparison is not working properly, as it will compare two very different datums and return that they are identical. Commenting this out for now. All input rasters wll experience a custom geotransformation until this is resolved.