Making lake pre-processing more flexible
Changes include allowing specification of the lake ID field to the reaches_with_lakes function, and giving the function the ability to detect that the input is a dicitonary rather than a feature class. This allows the user to send a dictionary that already specifies the mapping of reaches to waterbodies. That functionality can be useful if, for example, the lakes are too small to be placed on a reach using the normal spatial join method.