Resolving GROUNDWATER BASIN generation issue and Sample issue in Pro
For a while, the function to generate groundwater basin rasters has not completed successfully, resulting in a failure of the PROCESS_GEOGRID tool. After experimenting, we cleared all arcpy environments before running this function and it resolved the issue. Further, this update fixes an issue where the Sample tool fails in ArcGIS Pro. The resolution was to save inputs to GeoTiff on local disk (important) and save the output as a feature class (also essential).